requirements.txt
torch
gpytorch
ase
quippy-ase
pygps
sklearn
umap-learn
colorcet
locache
notebook
pandas